yebelo-trading-analytics/data-ingestion/requirements.txt
# Python dependencies for Data Ingestion Script

# Kafka client for Redpanda
kafka-python==2.0.2

# CSV parsing (built-in, but listed for documentation)
# No additional libraries needed for CSV